+++ app/models/unidom/certificate/certificating.rb
@@ -15,14 +15,26 @@ class Unidom::Certificate::Certificating < ActiveRecord::Base
   scope :certification_is, ->(certification) { where certification: certification }
 
   def self.certificate(certification: nil, certificated: nil, certificator: nil, opened_at: Time.now)
+    certificate! certification: certification, certificated: certificated, certificator: certificator, opened_at: opened_at
+  end
 
-    raise ArgumentError.new('certification is required.') if certification.blank?
-    raise ArgumentError.new('certificated is required.')  if certificated.blank?
+  def self.certificate!(certification: nil, certificated: nil, certificator: nil, opened_at: Time.now)
+    raise ArgumentError.new('The certification argument is required.') if certification.blank?
+    raise ArgumentError.new('The certificated argument is required.' ) if certificated.blank?
+    raise ArgumentError.new('The opened_at argument is required.'    ) if opened_at.blank?
 
-    self.certification_is(certification).certificated_is(certificated).valid_at.alive.first_or_create! opened_at: opened_at,
-      certificator_id:   (certificator.present? ? certificator.id         : Unidom::Common::NULL_UUID),
-      certificator_type: (certificator.present? ? certificator.class.name : '')
+    attributes = { opened_at: opened_at }
+    if certificator.present?
+      attributes[:certificator] = certificator
+    else
+      attributes[:certificator_id]   = Unidom::Common::NULL_UUID
+      attributes[:certificator_type] = ''
+    end
+    self.certification_is(certification).certificated_is(certificated).valid_at.alive.first_or_create! attributes
+  end
 
+  class << self
+    deprecate certificate: :certificate!, deprecator: ActiveSupport::Deprecation.new('2.0', 'unidom-certificate')
   end
 
 end