+++ app/models/unidom/certificate/concerns/as_certificated.rb
@@ -1,0 +1,11 @@
+module Unidom::Certificate::Concerns::AsCertificated
+
+  extend ActiveSupport::Concern
+
+  self.included do |includer|
+
+    has_many :certificatings, class_name: 'Unidom::Certificate::Certificating', as: :certificated
+
+  end
+
+end